+++ components/content/Content.css
@@ -1,0 +1,3 @@
+.headerTwo{
+    
+}
