+++ components/content/Content.jsx
@@ -1,5 +1,7 @@
 import React from 'react';
 
+import "./Content.css";
+
 export class Content extends React.Component {
 	render(){
 		return(
