+++ components/header/Header.css
@@ -1,0 +1,3 @@
+.headerOne{
+    
+}
