+++ components/header/Header.jsx
@@ -1,5 +1,7 @@
 import React from 'react';
 
+import "./Header.css";
+
 export class Header extends React.Component {
 	render() {
 		return(
